legacy/py-api/requirements-prod.txt
fastapi==0.131.0
uvicorn[standard]==0.41.0
sqlalchemy[asyncio]==2.0.46
asyncpg==0.31.0
psycopg[binary]==3.3.3
alembic==1.18.4
pydantic-settings==2.13.1
python-dotenv==1.2.1
argon2-cffi==25.1.0
email-validator==2.3.0
httpx==0.28.1
arq==0.27.0
redis==5.3.1
cryptography==46.0.5
curl-cffi==0.14.0
PyJWT[crypto]==2.11.0
webauthn>=2.0.0,<3.0.0
